requirements.txt
Flask==3.0.2
groq
pypdf
python-docx
pyyaml
pytesseract
pdf2image
Pillow
Jinja2
PyPDF2
gunicorn